// src/types/producers.types.ts
export const TIPO_DOCUMENTOS = [
  "dni_persona_fisica",
  "contrato_social",
  "dni_representante_legal",
  "comprobante_ISRC",
] as const;

export type TipoDocumento = (typeof TIPO_DOCUMENTOS)[number];

export interface Document {
  id_documento: string;
  ruta_archivo_documento: string;
  tipo_documento: TipoDocumento;
}
// ...
export interface GetDocumentsResponse {
  message: string;
  documentos: Document[];
}